requirements.txt
web3
pycoingecko
py-etherscan-api